Ability to customize localization
- `localized_string` uses a configurable class
to perform lookups
- FormBuilder.i18n_localizer is Formtastic::Localizer
by default; this maintains backwards compatability
- Updated documentation and generator accordingly
