+++ lib/formtastic.rb
@@ -12,6 +12,7 @@ module Formtastic
   autoload :I18n
   autoload :Inputs
   autoload :LocalizedString
+  autoload :Localizer
   autoload :Util
   
   # @private
